Move context menu into UI folder
