+++ trview.app/UI/ContextMenu.cpp
