+++ trview.app/UI/ContextMenu.h
